keep cleaning aeroelasticse
